Moved settings to top level as it was a pain to find.
