+++ src/app/settings/components/setting-content/settings-content.component.ts
