+++ src/app/settings/components/setting-header/settings-header.component.scss
