+++ src/app/settings/settings-panes/about/about-settings.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../mixins' as *;
+@use '../../../../mixins' as *;
 
 .overview {
   a {
