+++ src/app/settings/settings-panes/display/display-settings.component.html
