+++ src/app/settings/settings-panes/display/display-settings.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../mixins' as *;
+@use '../../../../mixins' as *;
 
 .themes {
   display: grid;
