+++ src/app/settings/settings-panes/localization/localization-settings.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../mixins' as *;
+@use '../../../../mixins' as *;
 
 .languages {
   display: flex;
