+++ src/app/settings/settings-panes/paired-devices/paired-devices-settings.component.html
