+++ src/app/settings/settings-panes/paired-devices/paired-devices-settings.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../mixins' as *;
+@use '../../../../mixins' as *;
 
 .buttons {
   display: flex;
