+++ src/app/settings/settings-panes/paired-devices/popups/change-endpoint/change-endpoint.component.html
