+++ src/app/settings/settings-panes/paired-devices/popups/change-endpoint/change-endpoint.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../../../mixins' as *;
+@use '../../../../../../mixins' as *;
 
 :host {
   background-color: var(--ui-bg-color-secondary);
