+++ src/app/settings/settings-panes/paired-devices/popups/change-endpoint/change-endpoint.component.ts
@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, Component, Inject, OnDestroy, OnInit, signal, WritableSignal }
   from '@angular/core';
-import { IconButtonComponent } from '../../../../../../components/icon-button/icon-button.component';
-import { IconInputComponent } from '../../../../../../components/icon-input/icon-input.component';
-import { PopupInstance } from '../../../../../../components/popup/popup.service';
-import { InstallService } from '../../../../../../services/install.service';
 import { Subject, takeUntil, tap } from 'rxjs';
+import { IconButtonComponent } from '../../../../../shared/components/icon-button/icon-button.component';
+import { IconInputComponent } from '../../../../../shared/components/icon-input/icon-input.component';
+import { PopupInstance } from '../../../../../shared/components/popup/popup.service';
+import { InstallService } from '../../../../../shared/services/install.service';
 
 @Component({
   selector: 'swc-change-endpoint',
