+++ src/app/settings/settings-panes/paired-devices/popups/pair-new-device/pair-new-device.component.html
