+++ src/app/settings/settings-panes/paired-devices/popups/pair-new-device/pair-new-device.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../../../mixins' as *;
+@use '../../../../../../mixins' as *;
 
 :host {
   background-color: var(--ui-bg-color-secondary);
