+++ src/app/settings/settings-panes/storage/storage-settings.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../mixins' as *;
+@use '../../../../mixins' as *;
 
 .state {
   background-color: var(--ui-bg-color-secondary);
