+++ src/app/settings/settings.component.html
@@ -19,5 +19,4 @@
       </swc-settings-content>
     }
   </div>
-
 }
