+++ src/app/settings/settings.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../mixins' as *;
+@use '../../mixins' as *;
 
 :host {
   --settings-padding: 2rem;
